+++ src/actions/clusterActions.js
@@ -232,12 +232,15 @@ export function clusterLoadDetails(clusterId) {
     var cluster;
     var clustersApi = new GiantSwarm.ClustersApi();
 
-    clustersApi.getClusterV5(scheme + ' ' + token, 'm0ckd').then(function(data) {
-      console.log('API called successfully. Returned data:');
-      console.log(data);
-    }, function(error) {
-      console.error(error);
-    });
+    clustersApi.getClusterV5(scheme + ' ' + token, 'm0ckd').then(
+      function(data) {
+        console.log('API called successfully. Returned data:');
+        console.log(data);
+      },
+      function(error) {
+        console.error(error);
+      }
+    );
 
     return clustersApi
       .getCluster(scheme + ' ' + token, clusterId)
@@ -263,7 +266,7 @@ export function clusterLoadDetails(clusterId) {
         throw error;
       });
   };
-} 
+}
 
 /**
  * Takes a clusterId and loads status for that cluster.
