+++ src/components/UI/theme.js
@@ -1,11 +1,13 @@
 const theme = {
   colors: {
+    background: '#234a61',
     shade1: '#1E4156',
     shade2: '#224A61',
     shade3: '#234D65',
     shade4: '#2F556A',
     shade5: '#32526A',
     shade6: '#3A5F7B',
+    shade7: '#30556a',
     white1: '#FFFFFF',
     white2: '#D7D7D7',
     white3: '#B9C1C8',
